requirements.txt
redis>=4.5.1
openai-whisper>=20231117
gradio>=4.0.0
psutil>=5.9.0
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic_settings
python-multipart>=0.0.6
google-generativeai>=0.3.0 
requests